refactor: Move `useUserAgents` hook from `assignables` plugin to `users`
