+++ plugins/leemons-plugin-assignables/frontend/src/components/Assignment/AssignStudents/hooks/index.js
@@ -1,4 +1,3 @@
-export { default as useUserAgents } from './useUserAgents';
 export { default as useTeacherClasses } from './useTeacherClasses';
 export { default as useTeacherClassesOfSubject } from './useTeacherClassesOfSubject';
 export { default as useGroupedClasses } from './useGroupedClasses';
