+++ plugins/leemons-plugin-assignables/frontend/src/components/Assignment/AssignStudents/hooks/useTeacherClasses.js
@@ -3,7 +3,7 @@ import _ from 'lodash';
 
 import { listTeacherClassesRequest } from '@academic-portfolio/request';
 import getCourseName from '@academic-portfolio/helpers/getCourseName';
-import useUserAgents from './useUserAgents';
+import useUserAgents from '@users/hooks/useUserAgents';
 
 // EN: Gets all the classes of the teacher
 // ES: Obtiene todas las clases del profesor
